+++ packages/pangea-sdk/pangea/__init__.py
@@ -1,6 +1,7 @@
-__version__ = "3.7.0"
+__version__ = "3.8.0beta2"
 
 from pangea.asyncio.request import PangeaRequestAsync
 from pangea.config import PangeaConfig
+from pangea.file_uploader import FileUploader
 from pangea.request import PangeaRequest
 from pangea.response import PangeaResponse