+++ packages/pangea-sdk/pangea/asyncio/services/__init__.py
@@ -4,4 +4,6 @@ from .embargo import EmbargoAsync
 from .file_scan import FileScanAsync
 from .intel import DomainIntelAsync, FileIntelAsync, IpIntelAsync, UrlIntelAsync, UserIntelAsync
 from .redact import RedactAsync
+from .sanitize import SanitizeAsync
+from .share import ShareAsync
 from .vault import VaultAsync