+++ packages/pangea-sdk/tests/integration/__init__.py
@@ -4,4 +4,6 @@ from .test_embargo import TestEmbargo
 from .test_file_scan import TestFileScan
 from .test_intel import TestDomainIntel, TestFileIntel, TestIPIntel, TestURLIntel, TestUserIntel
 from .test_redact import TestRedact
+from .test_sanitize import TestSanitize
+from .test_share import TestShare
 from .test_vault import TestVault