+++ CHANGELOG.md
@@ -7,8 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 3.8.0beta2 - 2024-03-20
+
 ### Added
 
+- Sanitize service support
+
+## 3.8.0beta1 - 2024-03-18
+
+### Added
+
+- Share service support
 - Audit /download_results endpoint support
 - AuthZ support